backend/requirements_api.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
bcrypt==4.1.2
PyJWT==2.8.0
email-validator==2.2.0